Fixed delete user logic in Manager, add Profile delete option

+++ src/main/java/com/bankapp/services/UserService.java
@@ -16,16 +16,12 @@ import org.springframework.transaction.annotation.Transactional;
 import com.bankapp.constants.Constants;
 import com.bankapp.encryption.RSAKeyPair;
 import com.bankapp.exceptions.EmailExistsException;
-import com.bankapp.models.Account;
 import com.bankapp.models.OneTimePassword;
 import com.bankapp.models.Role;
-import com.bankapp.models.Transaction;
 import com.bankapp.models.User;
 import com.bankapp.models.VerificationToken;
-import com.bankapp.repositories.AccountRepository;
 import com.bankapp.repositories.OTPRepository;
 import com.bankapp.repositories.RoleRepository;
-import com.bankapp.repositories.TransactionRepository;
 import com.bankapp.repositories.UserRepository;
 import com.bankapp.repositories.VerificationTokenRepository;
 
@@ -41,9 +37,6 @@ public class UserService implements IUserService, Constants {
     private UserRepository userRepository;
 
     @Autowired
-    private TransactionRepository transRepository;
-
-    @Autowired
     private OTPRepository oTPRepository;
 
     @Autowired
@@ -51,9 +44,6 @@ public class UserService implements IUserService, Constants {
 
     @Autowired
     private RoleRepository roleRepository;
-
-    @Autowired
-    private AccountRepository accountRepository;
 
     @Autowired
     private VerificationTokenRepository tokenRepository;
@@ -258,11 +248,22 @@ public class UserService implements IUserService, Constants {
             return SUCCESS;
         } catch (Exception e) {
             String logMessageFormat = "[Action=%s][Status=%s][User=%s, ErrorMessage=%s]";
-            String logMessage = String.format(logMessageFormat, "updateUser", ERROR, existingUserId,
-                    e.getMessage());
+            String logMessage = String.format(logMessageFormat, "updateUser", ERROR, existingUserId, e.getMessage());
             logger.info(logMessage);
 
             return ERROR;
+        }
+    }
+
+    @Override
+    public boolean markUserAsDeleted(String email) {
+        User user = userRepository.findByEmail(email);
+        if (user != null) {
+            user.setDeleted(true);
+            userRepository.save(user);
+            return true;
+        } else {
+            return false;
         }
     }
 
@@ -393,42 +394,29 @@ public class UserService implements IUserService, Constants {
     }
 
     @Override
-    public void deleteExternalUser(User user) {
-        Account account = accountRepository.findByUser(user);
+    public String deleteExternalUser(String email) {
+        try {
+            User user = userRepository.findByEmail(email);
 
-        VerificationToken verifyuser = tokenRepository.findByUser(user);
-        if (verifyuser != null) {
-            tokenRepository.delete(verifyuser);
-        }
-
-        if (account != null) {
-            List<Transaction> transactionfromAccount = transRepository.findByFromAccount(account);
-            List<Transaction> transactiontoAccount = transRepository.findByToAccount(account);
-
-            for (int i = 0; i < transactionfromAccount.size(); i++) {
-                transRepository.delete(transactionfromAccount.get(i));
+            if (user != null) {
+                userRepository.delete(user);
+            } else {
+                return ERR_ACCOUNT_NOT_EXISTS;
             }
 
-            for (int i = 0; i < transactiontoAccount.size(); i++) {
-                transRepository.delete(transactiontoAccount.get(i));
-            }
-
-            accountRepository.delete(account);
-
-            String logMessageFormat = "[Action=%s][Status=%s][User=%s, Action=%s]";
-            String logMessage = String.format(logMessageFormat, "deleteExternalUser", SUCCESS, user.getId(),
-                    "Deleted Account");
+            String logMessageFormat = "[Action=%s][Status=%s][User=%s]";
+            String logMessage = String.format(logMessageFormat, "deleteExternalUser", SUCCESS, email);
             logger.info(logMessage);
 
+            return SUCCESS;
+        } catch (Exception e) {
+            e.printStackTrace();
+            String logMessageFormat = "[Action=%s][Status=%s][User=%s, ErrorMessage=%s]";
+            String logMessage = String.format(logMessageFormat, "deleteExternalUser", ERROR, email, e.getMessage());
+            logger.error(logMessage);
+
+            return ERR_UNHANDLED;
         }
-
-        userRepository.delete(user);
-
-        String logMessageFormat = "[Action=%s][Status=%s][User=%s]";
-        String logMessage = String.format(logMessageFormat, "deleteExternalUser", SUCCESS, user.getId(),
-                "Deleted User");
-        logger.info(logMessage);
-
     }
 
 }
